Changed experience storage from integers to floats
This will make sure no experience is lost when the server has high
skill modifiers configured, or when low amounts of experience are being
shared.

+++ src/main/java/com/gmail/nossr50/skills/SkillManager.java
@@ -38,7 +38,7 @@ public abstract class SkillManager {
         return activationChance;
     }
 
-    public void applyXpGain(int xp) {
+    public void applyXpGain(float xp) {
         mcMMOPlayer.beginXpGain(skill, xp);
     }
 }
